+++ comp2013/src/main/java/comp2013/MyFrame.java
@@ -156,7 +156,7 @@ public class MyFrame extends JPanel implements KeyListener
 					left = false;
 					right = false;
 
-					newImgSnakeHead = (BufferedImage) GameUtil.rotateImage(IMG_SNAKE_HEAD, -90);
+					newImgSnakeHead = (BufferedImage) ImageUtil.rotateImage(IMG_SNAKE_HEAD, -90);
 				}
 				break;
 
@@ -168,7 +168,7 @@ public class MyFrame extends JPanel implements KeyListener
 					left = false;
 					right = false;
 
-					newImgSnakeHead = (BufferedImage) GameUtil.rotateImage(IMG_SNAKE_HEAD, 90);
+					newImgSnakeHead = (BufferedImage) ImageUtil.rotateImage(IMG_SNAKE_HEAD, 90);
 				}
 				break;
 
@@ -180,7 +180,7 @@ public class MyFrame extends JPanel implements KeyListener
 					left = true;
 					right = false;
 
-					newImgSnakeHead = (BufferedImage) GameUtil.rotateImage(IMG_SNAKE_HEAD, -180);
+					newImgSnakeHead = (BufferedImage) ImageUtil.rotateImage(IMG_SNAKE_HEAD, -180);
 
 				}
 				break;